+++ src/pypex/base/shape.py
@@ -47,7 +47,7 @@ class Shape2D(metaclass=ABCMeta):
 
     @hull.setter
     def hull(self, hull):
-        self._hull = hull
+        self._hull = np.array(hull)
 
     @abstractmethod
     def intersects(self, shape, **kwargs):
@@ -76,8 +76,8 @@ class Shape2D(metaclass=ABCMeta):
     def to_Points(self):
         return np.array([Point(*point) for point in self.hull])
 
-    # redefine to_Points also to lowercase
-    to_points = to_Points
+    def to_array(self):
+        return np.array(self._hull)
 
     @staticmethod
     def validity_check(hull, length, _raise=True):
@@ -94,3 +94,13 @@ class Shape2D(metaclass=ABCMeta):
     @classmethod
     def polygon_validity_check(cls, hull, _raise=True):
         return cls.validity_check(hull, 1, _raise)
+
+    @property
+    def transpose(self):
+        """
+        Transpose self._hull and return numpy.array.
+        :return: numpy.array
+        """
+        return self._hull.T
+
+    T = transpose